+++ src/App.vue
@@ -1,23 +1,11 @@
 <template>
-  <tyh-tag v-show="show1" isClose @close="show1 = false">标签一</tyh-tag>
-  <tyh-tag v-show="show2" type="primary" isClose @close="show2 = false">
-    标签二
-  </tyh-tag>
-  <tyh-tag v-show="show3" type="success" isClose @close="show3 = false">
-    标签三
-  </tyh-tag>
-  <tyh-tag v-show="show4" type="danger" isClose @close="show4 = false">
-    标签四
-  </tyh-tag>
-  <tyh-tag v-show="show5" type="warning" isClose @close="show5 = false">
-  </tyh-tag>
+  <tyh-tagging>标注1</tyh-tagging>
+  <tyh-tagging type="">标注2</tyh-tagging>
+  <tyh-tagging type="success">标注3</tyh-tagging>
+  <tyh-tagging type="danger">标注4</tyh-tagging>
+  <tyh-tagging type="warning">标注5</tyh-tagging>
 </template>
 
-<script setup>
-import { ref } from 'vue'
-const show1 = ref(true)
-const show2 = ref(true)
-const show3 = ref(true)
-const show4 = ref(true)
-const show5 = ref(true)
-</script>
+<script setup></script>
+
+<style scoped></style>