+++ src/packages/tyhUi/packages/tagging/index.js
@@ -1,4 +1,4 @@
-import T from './TyhTagging.vue'
+import T from './src/TyhTagging.vue'
 
 T.install = _ => _.component('TyhTagging', T)